another fix

+++ src/main/java/com/example/bandservice/repository/BandRepository.java
@@ -9,13 +9,16 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface BandRepository extends CassandraRepository<Band,Integer> {
+public interface BandRepository extends CassandraRepository<Band, Long> {
     @Query("select * from band where name=:name allow filtering")
     Band findByName(@Param("name") String name);
+
     @Query("select * from band")
     List<Band> findAllBands();
+
     @Query("select * from band where id=:id allow filtering")
-    Band getBandById(@Param("id")int id);
+    Band getBandById(@Param("id") Long id);
+
     @Query("update band set name=:name where id=:id")
-    Band update(@Param("id")int id, @Param("name")String name);
+    Band update(@Param("id") Long id, @Param("name") String name);
 }
